requirements.txt
asgiref==3.8.1
# backports.zoneinfo==0.2.1
Django==4.2.14
django-crontab==0.7.1
django-mail-queue==3.2.5
pillow==10.4.0
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.1
